refactor(frontend): use RadioGroup instead of Select for extraction aspect ratio selection
