+++ apps/frontend/src/widgets/SettingsPanel/TabResize/sections/extract/ExtractContent/RadioGroupExtractRatio.module.css
@@ -1,0 +1,7 @@
+.title {
+  font-weight: 600;
+}
+
+.item {
+  text-transform: capitalize;
+}
