--- apps/frontend/src/widgets/SettingsPanel/TabResize/sections/extract/ExtractContent/SelectExtractAspectRatio.module.css
@@ -1,3 +1,0 @@
-.trigger {
-  max-width: 100px;
-}
